+++ packages/ui-components/docs/changelog.md
@@ -2,6 +2,11 @@
 title: Changelog
 ---
 
+## NEXT
+
+-   Added `onKeyDown` prop to NumericInput to bubble up keydown events
+-   Added the `onChange` property to the `onChange` callback dependency
+
 ## 1.4.3
 
 -   Pinned fortawesome to `~6.4.0`